+++ projects/ngx-rut-v2/src/lib/directives/rut-value-accessor.directive.spec.ts
@@ -3,6 +3,7 @@ import { TestBed } from '@angular/core/testing';
 import { FormControl, ReactiveFormsModule } from '@angular/forms';
 import { By } from '@angular/platform-browser';
 import { RutValueAccessor } from './rut-value-accessor.directive';
+import { describe, expect, it } from 'vitest';
 
 @Component({
   standalone: true,